backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
httpx
python-multipart
python-dotenv
pydantic[email]
pydantic-settings
